Remove invalid assertion from Kernel.php
Previously added this assertion to silence PHP CS. This assertion is
causing the assertion to fail, as the object under test is not
Traversable. Removed the assertion to fix the issue. The CS issue was
already fixed correctly by adding a doc block with a return type
annotation.

Note to self, next time RTFM