Add a webhook to receive session verification
After the user finish the passport verification process, ReadId server must be 
configured to send a push notification to 'readid_webhook' route.

The route requires a secret token configured in  _APP_WEBHOOK_TOKEN_ 
parameter to avoid someone tries to do a manual assertion.

The controller dispatch a ConfirmPendingSessionCommand to persist the
session id in the PendingSession.