Fix sso redirect url from gssp-bundle
The sso_redirect url was used from the Authentication service
in the GSSP-bundle. Because only certain classes are use from the
GSSP-bundle and not all the bundle related stuff such as routes,
contollers etc. the missing route resulted in an exception.

Now the url is generated with the helper method of the autentication
controller in ReadID itself.